+++ packages/editor/src/plugins/index.ts
@@ -1,2 +1,3 @@
 export * from './selection-change';
 export * from './plugin';
+export * from './undo';