+++ packages/editor/src/plugins/undo.ts
@@ -1,0 +1,41 @@
+import { useEffect } from 'react';
+import { redo, undo, history, undoHistory } from '../constants/atom';
+import { useEventListener } from '../hooks';
+import { usePlugin } from './plugin';
+
+export function Unstable_UndoPlugin() {
+  const { wrapperRef } = usePlugin();
+
+  useEventListener(
+    'keydown',
+    (e) => {
+      const buttonKey = /(Mac)/i.test(navigator.platform)
+        ? 'metaKey'
+        : 'ctrlKey';
+      if (e[buttonKey] && e.key === 'z' && !e.shiftKey) {
+        if (history.length > 0) {
+          undo();
+          e.preventDefault();
+        }
+      }
+    },
+    wrapperRef,
+  );
+
+  useEventListener(
+    'keydown',
+    (e) => {
+      const buttonKey = /(Mac)/i.test(navigator.platform)
+        ? 'metaKey'
+        : 'ctrlKey';
+      if (e[buttonKey] && e.shiftKey && e.key === 'z') {
+        if (undoHistory.length > 0) {
+          redo();
+          e.preventDefault();
+        }
+      }
+    },
+    wrapperRef,
+  );
+  return null;
+}