+++ packages/editor/src/utils/insert.ts
@@ -10,3 +10,13 @@ export const insert = <T>(
   // part of the array after the specified index
   ...arr.slice(index),
 ];
+
+export const insertArray = <T>(
+  arr: Array<T>,
+  index: number,
+  newItems: Array<T>,
+) => [
+  ...arr.slice(0, index),
+  ...newItems,
+  ...arr.slice(index + newItems.length),
+];